Modification 01/06

+++ src/Controller/AccountController.php
@@ -5,6 +5,9 @@ namespace App\Controller;
 
 use App\Entity\Server;
 use App\Form\Type\ServerType;
+use App\Repository\ServerRepository;
+use App\Service\NameGenerator;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -13,15 +16,39 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class AccountController extends AbstractController
 {
     /**
-     * @Route ("/account", name="page_account")
+     * @var NameGenerator $nameGenerator
+     */
+    private $nameGenerator;
+
+    /**
+     * @var ServerRepository $serverRepository
+     */
+    private $serverRepository;
+
+    /**
+     * AccountController constructor.
+     * @param NameGenerator $nameGenerator
+     * @param ServerRepository $serverRepository
+     */
+    public function __construct(NameGenerator $nameGenerator, ServerRepository $serverRepository)
+    {
+        $this->nameGenerator = $nameGenerator;
+        $this->serverRepository = $serverRepository;
+    }
+
+
+    /**
+     * @Route ("/account", name="account_dashboard")
      */
     public function account(): Response
     {
-        return $this->render('account/account_dashboard.html.twig');
+        $server = $this->serverRepository->findBy([], ['id'=> 'ASC']);
+
+        return $this->render('account/account_dashboard.html.twig', ['server_list'=> $server]);
     }
 
     /**
-     * @Route ("/account/profil", name="page_profile")
+     * @Route ("/account/profil", name="account_profile")
      */
     public function profile(): Response
     {
@@ -29,32 +56,37 @@ class AccountController extends AbstractController
     }
 
     /**
-     * @Route ("/account/new-server", name="page_new-server")
+     * @Route ("/account/new-server", name="account_new-server")
      */
     public function create(Request $request): Response
     {
         $server = new Server();
 
-        $form = $this->createForm(ServerType::class);
+        $form = $this->createForm(ServerType::class, $server);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
 
+            $this->nameGenerator->generate($server);
+
             $em->persist($server);
             $em->flush();
+
+            $url = $this->generateUrl('page_server', ['id'=> $server->getId()] );
+            return new RedirectResponse($url);
         }
 
-        return $this->render('account/account_new-server.html.twig', ['form' => $form->createView()]);
+        return $this->render('account_new-server.html.twig', ['form' => $form->createView()]);
     }
 
     /**
-     * @Route ("/account/{server.id}", name="page_server")
+     * @Route ("/account/{id}", name="account_server")
      */
-    public function server(): Response
+    public function server(int $id): Response
     {
-        $server = $this->ServerRepository->findOneById($id);
+        $server = $this->serverRepository->findOneById($id);
 
         if (null === $server) {
             throw $this->createNotFoundException('Serveur introuvable');
@@ -64,17 +96,9 @@ class AccountController extends AbstractController
     }
 
     /**
-     * @Route ("/account/{server.id}/reboot", name="page_server_reboot")
+     * @Route ("/account/{id}/delete", name="page_server_delete")
      */
-    /*public function serverReboot(): Response
-    {
-        return $this-render();
-    }
-
-    /**
-     * @Route ("/account/{server.id}/delete", name="page_server_delete")
-     */
-    public function serverDelete(): Response
+    public function serverDelete(Request $request, int $id): Response
     {
         $repository = $this->getDoctrine()->getRepository(Server::class);
 
@@ -84,19 +108,78 @@ class AccountController extends AbstractController
             throw $this->createNotFoundException('Serveur introuvable');
         }
 
-        $em = $this->setDoctrine()->getManager();
+        $form = $this->createForm(ServerType::class, $server);
+        $form->handleRequest($request);
 
-        $em->remove($server);
-        $em->flush();
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($server);
+            $em->flush();
+        }
 
-        return $this->render('account/server/delete.html.twig');
+        return $this->redirectToRoute('account_dashboard');
     }
 
     /**
-     * @Route ("/account/api/{server.id}/ready", name="page_server_ready")
+     * @Route ("/account/{id}/reboot", name="server_reboot")
      */
-    /*public function serverReboot(): Response
+    public function Reboot(int $id): Response
     {
-        return $this->render();
-    }*/
+        /**
+         * @var Server $server
+         */
+
+        $server = $this->serverRepository->findOneById($id);
+        $server->setState(Server::STATE_STOPPED);
+        $em = $this->getDoctrine()->getManager();
+
+        $em->persist($server);
+        $em->flush();
+
+        $this->addFlash('success', 'Votre serveur redémarre');
+        return $this->redirectToRoute('account_server', $id);
+    }
+
+    /**
+     * @Route ("/account/{id}/reset", name="server_reset")
+     */
+    public function reset(int $id): Response
+    {
+        /**
+         * @var Server $server
+         */
+
+        $server = $this->serverRepository->findOneById($id);
+        $server->setState(Server::STATE_PENDING);
+        $em = $this->getDoctrine()->getManager();
+
+        $em->persist($server);
+        $em->flush();
+
+        $this->addFlash('success', 'Votre serveur est réinstallé');
+        return $this->redirectToRoute('account_server', $id);
+    }
+
+
+    /**
+     * @Route ("/account/api/{id}/ready", name="page_server-ready")
+     */
+    public function Ready(): Response
+    {
+        /**
+         * @var Server $server
+         */
+
+        $server = $this->serverRepository->findOneById($id);
+        $server->setState(Server::STATE_READY);
+        $em = $this->getDoctrine()->getManager();
+
+        $em->persist($server);
+        $em->flush();
+
+        $this->addFlash('success', 'Votre serveur est prêt');
+        return $this->redirectToRoute('account_server');
+
+        //TODO : envoyer mail au client
+    }
 }
